Add hfsplus command 'setfinderflags' to set finder info flags
This is useful for building dmg images in userspace
where having a nice volume icon requires setting the
kHasCustomIcon flag on the root directory of the
filesystem.

Support the following flags;
* HasCustomIcon
* IsStationary (Only for files)
* NameLocked
* HasBundle (Only for files)
* IsInvisible
* IsAlias (Only for files)

Each flag is prefixed with either '+' or '-' to indicate
whether the flag should be set or cleared respectively.

Example usage:

`hfsplus mydisk.hfsplus.img setfinderflags / +HasCustomIcon`

Ref: https://developer.apple.com/library/archive/technotes/tn/tn1150.html#FinderInfo